+++ AssetsFlow.Server/AssetsFlow.Web.Services/Models/Journal/TradeCompositeModel.cs
@@ -1,5 +1,6 @@
 ﻿using HsR.Common.Extenders;
 using HsR.Journal.Entities;
+using HsR.Journal.TradeAnalytics;
 using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 
@@ -25,6 +26,8 @@ namespace AssetsFlowWeb.Services.Models.Journal
         public DateTime? OpenedAt { get; set; }
         public DateTime? ClosedAt { get; set; }
         public DateTime? IdeaDate { get; set; }
+
+        public TradeAnalyticsSummary Analytics { get; set; }
 
         public override string ToString() => $"Id={Id}, IsAnyContentMissing={IsAnyContentMissing}, Status={Status}";
     }