+++ AssetsFlow.Server/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Map date stamps on composite dto, removed double entries on composite models
 - DenyActivation flag added to content edit method
 - namespaces foxes
+- return trade analytics in status
 
 ## [0.9.220-b] - 2025-07-26
 - User Service with gRPC + host + unittests